Larger DescriptionIndex (#3715)
- Change DescriptionIndex data type to uint16_t, because we're running out of IDs for Building Descriptions
- Fix usages of DescriptionIndex
- Rename function get_resource to resource_index for consistency.

+++ src/logic/widelands.h
@@ -59,9 +59,9 @@ using TeamNumber = uint8_t;
  * EditorCategory, BuildingDescr, ImmovableDescr, ShipDescr, TribeDescr
  * and others.
  */
-using DescriptionIndex = uint8_t;
+using DescriptionIndex = uint16_t;
 
-constexpr DescriptionIndex INVALID_INDEX = std::numeric_limits<uint8_t>::max();
+constexpr DescriptionIndex INVALID_INDEX = std::numeric_limits<uint16_t>::max();
 constexpr DescriptionIndex kInvalidWare = INVALID_INDEX - 1;
 constexpr DescriptionIndex kNoResource = INVALID_INDEX - 1;
